fixed modalities, and online status display, contact region arragnemnt in therapist item
