+++ lib/features/therapists/presentation/widgets/therapists_table.dart
@@ -10,6 +10,7 @@ import 'package:tranquil_admin_portal/core/utils/extensions/date_time_extension.
 import 'package:tranquil_admin_portal/core/utils/helpers/size_helpers.dart';
 import 'package:tranquil_admin_portal/features/therapists/domain/entities/therapist.dart';
 import 'package:tranquil_admin_portal/features/therapists/presentation/controllers/therapists_controller.dart';
+import 'package:tranquil_admin_portal/features/therapists/presentation/widgets/therapist_item.dart';
 
 class TherapistsTable extends StatefulWidget {
   const TherapistsTable({super.key});
@@ -32,44 +33,6 @@ class _TherapistsTableState extends State<TherapistsTable> {
     'LAST ACTIVE',
     'ACTIONS'
   ];
-
-  // var therapists = <Map<String, dynamic>>[
-  //   {
-  //     'id': 1,
-  //     'name': 'Elyse Lindell, MD',
-  //     'contact': {
-  //       'email': 'elyselindell@gmail.com',
-  //       'phone': '+44 20 8763 1129',
-  //     },
-  //     'location':{
-  //       'longitude':,
-  //       'latitude':,
-  //       'name": '',
-  //     }
-  //     'type': {'company_name': 'Private practice', 'type': 'Solo practitioner'},
-  //     'earnings': {'amount_earned': '45094', 'completed_sessions': '33'},
-  //     'date_joined': 'Sep 27, 2023',
-  //     'last_active': 'ONLINE',
-  //     'selected': false
-  //   },
-  //   {
-  //     'id': 2,
-  //     'name': 'Sebastian Thoren, MD',
-  //     'contact': {
-  //       'email': 'sebastainthoren@harleypsych.com',
-  //       'phone': '+1 555 555 1129',
-  //       'location': 'USA'
-  //     },
-  //     'type': {
-  //       'company_name': 'Agency-based therapy',
-  //       'type': 'Harley Psychiatrists'
-  //     },
-  //     'earnings': {'amount_earned': '59405', 'completed_sessions': '29'},
-  //     'date_joined': 'Today, 12:07pm',
-  //     'last_active': 'ONLINE',
-  //     'selected': false
-  //   },
-  // ];
 
   void toggleSelection(int id, List<Therapist> therapists) {
     setState(() {
@@ -223,184 +186,4 @@ class _TherapistsTableState extends State<TherapistsTable> {
   }
 }
 
-class TherapistItem extends StatelessWidget {
-  const TherapistItem(
-      {super.key,
-      required this.item,
-      required this.onCheckboxChanged,
-      required this.therapistsController});
 
-  final Therapist item;
-  final TherapistsController therapistsController;
-  final ValueChanged<int> onCheckboxChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    final GlobalKey actionKey = GlobalKey();
-
-    return Column(
-      children: [
-        Table(
-          defaultVerticalAlignment: TableCellVerticalAlignment.middle,
-          columnWidths: const {
-            0: FixedColumnWidth(40), // Checkbox column
-            1: FixedColumnWidth(50), // S/N column
-            2: FlexColumnWidth(), // PROFILE column
-            3: FlexColumnWidth(), // CONTACT column
-            4: FlexColumnWidth(), // TYPE column
-            5: FlexColumnWidth(), // EARNINGS column
-            6: FlexColumnWidth(), // DATE JOINED column
-            7: FlexColumnWidth(), // LAST ACTIVE column
-            8: FixedColumnWidth(80), // ACTIONS column
-          },
-          children: [
-            TableRow(children: [
-              Checkbox(
-                value: false,
-                onChanged: (_) {},
-              ),
-              CustomText(
-                text: item.id.toString().toUpperCase(),
-                size: 12,
-              ),
-              CustomText(
-                text: "${item.fName} ${item.lName}",
-                size: 12,
-              ),
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Wrap(
-                    spacing: 4,
-                    direction: Axis.horizontal,
-                    crossAxisAlignment: WrapCrossAlignment.center,
-                    children: [
-                      SvgPicture.asset(SvgPaths.emailIcon),
-                      CustomText(
-                        text: item.email,
-                        size: 12,
-                      ),
-                    ],
-                  ),
-                  SizedBox(height: 2),
-                  Wrap(
-                    spacing: 4,
-                    direction: Axis.horizontal,
-                    crossAxisAlignment: WrapCrossAlignment.center,
-                    children: [
-                      SvgPicture.asset(SvgPaths.phoneIcon),
-                      CustomText(
-                        text: item.phone,
-                        size: 12,
-                      ),
-                    ],
-                  )
-                ],
-              ),
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  CircleAvatar(
-                    maxRadius: 16,
-                    child: item.avatarUrl == null ? Placeholder() : Image.network(item.avatarUrl!),
-                  ),
-                  SizedBox(width: 4),
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      CustomText(
-                        text: "null",
-                        size: 12,
-                      ),
-                      CustomText(
-                        text: "null",
-                        size: 12,
-                      ),
-                    ],
-                  )
-                ],
-              ),
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  CustomText(
-                    text: "\$${item.earnings?.netIncome}",
-                    size: 12,
-                    weight: FontWeight.bold,
-                  ),
-                  CustomText(
-                    text: '${item.completedSessions} sessions',
-                    size: 12,
-                  ),
-                ],
-              ),
-              CustomText(
-                text: item.createdAt?.toLocal().formatDateTime,
-                size: 12,
-              ),
-              CustomText(
-                text: item.updatedAt?.toLocal().formatDateTime,
-                size: 12,
-              ),
-              GestureDetector(
-                key: actionKey, // Assign the key here
-                onTap: () => displayActionPopUp(
-                    context, therapistsController, actionKey, item.id!),
-                child: SvgPicture.asset(SvgPaths.moreIcon),
-              )
-            ])
-          ],
-        ),
-        Divider()
-      ],
-    );
-  }
-
-  displayActionPopUp(
-      BuildContext context,
-      TherapistsController therapistsController,
-      GlobalKey key,
-      int therapistId) {
-    // Get the RenderBox and its position
-    final RenderBox renderBox =
-        key.currentContext!.findRenderObject() as RenderBox;
-    final Offset offset = renderBox.localToGlobal(Offset.zero);
-    final Size size = renderBox.size;
-
-    // Show the menu at the calculated position
-    return showMenu(
-      color: Colors.white, // Set your preferred color
-      context: context,
-      position: RelativeRect.fromLTRB(
-        offset.dx, // X-coordinate of the GestureDetector
-        offset.dy + size.height, // Y-coordinate + height of the item
-        offset.dx + size.width, // Width of the GestureDetector
-        0, // No margin at the bottom
-      ),
-      items: [
-        PopupMenuItem(
-          onTap: () async =>
-              await therapistsController.approveTherapist(therapistId),
-          value: 'approve',
-          child: const Text('Approve'),
-        ),
-        const PopupMenuItem(
-          value: 'block',
-          child: Text('Block'),
-        ),
-        const PopupMenuItem(
-          value: 'view',
-          child: Text('View'),
-        ),
-        const PopupMenuItem(
-          value: 'pay',
-          child: Text('Pay'),
-        ),
-        const PopupMenuItem(
-          value: 'delete',
-          child: Text('Delete'),
-        ),
-      ],
-    );
-  }
-}
